refactor(runtime): extract tool_call module from agent_loop
The largest single extraction in this PR — the tool-call execution
path (~1,090 LOC) moves into `agent_loop/tool_call.rs`. Covers four
distinct responsibilities:

**1. Tool-call lifecycle structs (`pub(super)` + `pub(super)` fields
because the main loop constructs `ToolExecutionContext` literally and
calls accessor / mutator methods on `StagedToolUseTurn` directly):**

- `ToolResultOutcomeSummary` — per-iteration hard/success counter
  used by the consecutive-failure abort threshold.
- `StagedToolUseTurn` — the #2381 structural fix: stages assistant +
  tool_result blocks locally and commits atomically, so a `?` exit
  between LLM call and tool-result push can't half-commit the
  session and leave orphaned `tool_use` blocks.
- `ExecutedToolCall` — the `(ToolResult, final_content)` pair
  returned by `execute_single_tool_call` (caller reads both).
- `ToolExecutionContext<'a>` — the 30-field parameter pack for
  `execute_single_tool_call_inner`.

**2. Per-tool dispatch helpers:**

- `tool_use_blocks_from_calls` — `ToolCall` → `ContentBlock::ToolUse`
  list builder for the assistant message.
- `sanitize_tool_label` + `record_tool_call_metric` — bounded
  metric-label sanitizer + the #3495 observability counter.
- `append_tool_result_guidance_blocks` — appends system-text blocks
  to the user message when denials, modify-and-retry feedback,
  parameter errors, or non-parameter errors appear in the turn's
  tool results (separates LLM-recoverable from hard errors).
- `update_consecutive_hard_failures` — bumps or resets the
  consecutive-failure counter based on the outcome summary.

**3. The execution path itself:**

- `stage_tool_use_turn` — builds the `StagedToolUseTurn` from the
  raw `tool_calls` slice.
- `execute_single_tool_call` (instrumented wrapper) +
  `execute_single_tool_call_inner` — the ~280-LOC per-tool
  dispatcher (timeout, interrupt, approval, sandbox, taint, hooks).
  The wrapper guarantees `record_tool_call_metric` fires on every
  return path (Ok + Err).

**4. Post-LLM tool-result synthesis:**

- `append_skipped_tool_results` — fabricates synthetic
  "skipped" results for tool_use ids the per-tool loop never
  reached (mid-turn signal, interrupt).
- `handle_mid_turn_signal` — routes mid-turn signals (approval
  resolution, interrupt) through the right side-channel.
- `session_owns_pending_approval` — TOCTOU guard for
  `apply_approval_resolution_signal` so a stale signal from
  another session can't patch the wrong `WaitingApproval` block.
- `finalize_tool_use_results` — runs the guidance-block append
  and the per-iteration outcome summary.
- `apply_approval_resolution_signal` — patches `WaitingApproval`
  blocks in both `session.messages` and the in-flight messages
  slice when an approval comes back mid-turn.

Test imports updated: `mod tests` picks up `finalize_tool_use_results`,
`record_tool_call_metric`, and `StagedToolUseTurn` via the existing
test-only `use super::tool_call::{...}` pattern. The stranded
`/// Patches WaitingApproval blocks…` doc-comment that the deletion
range left above `fire_hook_best_effort` is removed in the same
commit (fixes the `clippy::empty-line-after-doc-comments` lint).

Behavior-preserving. mod.rs: 10,284 -> 9,197 LOC (-1,087).
cargo check + cargo clippy --lib + cargo check --tests all clean.

Refs #3710
